Changed minimum supported MC version to 1.16
