+++ src/main/java/net/coreprotect/paper/PaperAdapter.java
@@ -41,10 +41,8 @@ public class PaperAdapter implements PaperInterface {
             case PAPER_V1_13:
             case PAPER_V1_14:
             case PAPER_V1_15:
-                PaperAdapter.ADAPTER = new PaperHandler();
-                break;
             case PAPER_V1_16:
-                PaperAdapter.ADAPTER = new Paper_v1_16();
+                PaperAdapter.ADAPTER = new PaperHandler();
                 break;
             case PAPER_V1_17:
             case PAPER_V1_18:
