+++ src/main/java/net/coreprotect/paper/PaperHandler.java
@@ -3,6 +3,8 @@ package net.coreprotect.paper;
 import org.bukkit.Location;
 import org.bukkit.Server;
 import org.bukkit.entity.Entity;
+import org.bukkit.inventory.Inventory;
+import org.bukkit.inventory.InventoryHolder;
 
 public class PaperHandler extends PaperAdapter implements PaperInterface {
 
@@ -14,6 +16,11 @@ public class PaperHandler extends PaperAdapter implements PaperInterface {
     @Override
     public void teleportAsync(Entity entity, Location location) {
         entity.teleportAsync(location);
+    }
+
+    @Override
+    public InventoryHolder getHolder(Inventory holder, boolean useSnapshot) {
+        return holder.getHolder(useSnapshot);
     }
 
 }
