+++ src/main/java/net/coreprotect/paper/Paper_v1_17.java
@@ -4,7 +4,7 @@ import org.bukkit.block.Sign;
 
 import net.kyori.adventure.text.serializer.legacy.LegacyComponentSerializer;
 
-public class Paper_v1_17 extends Paper_v1_16 implements PaperInterface {
+public class Paper_v1_17 extends PaperHandler implements PaperInterface {
 
     @Override
     public String getLine(Sign sign, int line) {
